Grammar definition is about halfway done.
However, I still need to figure out how to define some of the grammar
dynamically, so it doesn't necessitate an update to the extension every
time a new keyword or function is added. This might be possible through
the language server, but for now, I think finishing the static grammar
will be enough to release the first version.